lexeme files
