+++ C/parser/Parser__IMPL__.inc
@@ -25,7 +25,7 @@
 
 #include "compilation/Compilation.h"
 #include "syntax/SyntaxFacts.h"
-#include "syntax/SyntaxLexemes.h"
+#include "syntax/SyntaxLexeme_ALL.h"
 #include "syntax/SyntaxNodes.h"
 #include "syntax/SyntaxToken.h"
 #include "syntax/SyntaxUtilities.h"
