+++ C/stdlib-support/StdLibInterceptor.cpp
@@ -20,7 +20,7 @@
 
 #include "StdLibInterceptor.h"
 
-#include "syntax/SyntaxLexemes.h"
+#include "syntax/SyntaxLexeme_ALL.h"
 #include "syntax/SyntaxNodes.h"
 
 #include <iostream>
