+++ C/tests/ParserTest_0000_0999.cpp
@@ -24,7 +24,7 @@
 
 #include "parser/Parser.h"
 #include "parser/Unparser.h"
-#include "syntax/SyntaxLexemes.h"
+#include "syntax/SyntaxLexeme_ALL.h"
 #include "syntax/SyntaxNodes.h"
 
 #include <algorithm>
