+++ C/tests/Test.cpp
@@ -27,7 +27,7 @@
 #include "symbols/Symbol.h"
 #include "parser/Unparser.h"
 #include "symbols/Symbol_ALL.h"
-#include "syntax/SyntaxLexemes.h"
+#include "syntax/SyntaxLexeme_ALL.h"
 #include "syntax/SyntaxNamePrinter.h"
 #include "syntax/SyntaxNodes.h"
